Rename CompletedHomeworksViewBloc file.
